requirements.txt
RPi.GPIO==0.7.0
aiocoap==0.3
LinkHeader==0.4.3
